Update GPyTorchUtils
Updated GPyTorchUtils by moving two data processing functions
torch_one_hour_data_split and torch_one_hour_target_split which are used
for two experiements.

The first (torch_one_hour_data_split) being creating a shifting window
of data where each column is a 24 sample chunk (a day) of data with
another chunk of a next day with an offset. It is assumed that the data
set is a multiple of 24.

The second (torch_one_hour_target_split) takes a offset between 0 and 23
(denoted as k) and gets all samples of 24n + k from the provided data
set. It is assumed that the data set is a multiple of 24.